исправление ошибок

+++ lab_1_classify_by_unigrams/main.py
@@ -60,7 +60,7 @@ def calculate_frequencies(tokens: list[str] | None) -> dict[str, float] | None:
         freq_dictionary[letter] += 1
 
     for letter in freq_dictionary:
-        freq_dictionary[letter] = int(freq_dictionary[letter]) / length
+        freq_dictionary[letter] = float(freq_dictionary[letter]) / length
     return freq_dictionary
